test: separate TestInstancesReview from test file for easier access
