+++ app/src/androidTest/java/com/github/se/cyrcle/model/review/ReviewViewModelTest.kt
@@ -22,16 +22,6 @@ import org.junit.Assert.fail
 import org.junit.Before
 import org.junit.Test
 
-object TestInstancesReview {
-  val review1 =
-      Review(uid = "1", owner = "user1", text = "Great parking!", rating = 5.0, parking = "P1")
-  val review2 =
-      Review(uid = "2", owner = "user2", text = "Okay parking.", rating = 3.0, parking = "P2")
-  val review3 =
-      Review(uid = "3", owner = "user1", text = "Bad Parking.", rating = 1.0, parking = "P3")
-  val review4 =
-      Review(uid = "4", owner = "user3", text = "New Review.", rating = 4.5, parking = "P4")
-}
 
 class ReviewViewModelTest {
 
@@ -235,7 +225,7 @@ class ReviewViewModelTest {
         })
 
     // Wait for Firestore to respond, or timeout after 5 seconds
-    val completed = countDownLatch.await(5, TimeUnit.SECONDS)
+    val completed = countDownLatch.await(1, TimeUnit.SECONDS)
 
     // Assert that the test completed within the timeout
     assertTrue("Test timed out", completed)
